refactor

+++ src/main/java/mob/code/supermarket/bean/BuyItem.java
@@ -1,6 +1,7 @@
 package mob.code.supermarket.bean;
 
 import lombok.AllArgsConstructor;
+import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 
@@ -11,12 +12,12 @@ public class BuyItem {
 
     private String barcode;
 
-    private Integer count;
+    private double count;
 
     public static BuyItem from(String str) {
         String[] split = str.split("-");
         if (split.length == 2) {
-            return new BuyItem(split[0], Integer.parseInt(split[1]));
+            return new BuyItem(split[0], Double.parseDouble(split[1]));
         }
         return new BuyItem(str, 1);
     }
